refactor(web): reorganize task details component structure
- Moved the TaskDetailsComponent to a new file for better organization and clarity.
- Updated import paths in related components to reflect the new structure.
- Simplified the MainLayout component by removing unnecessary fragments around children.
