+++ apps/web/app/[locale]/(main)/layout.tsx
@@ -5,7 +5,7 @@ export default function MainLayout({ children }: { children: React.ReactNode }) 
 	return (
 		<>
 			<AppState />
-			<>{children}</>
+			{children}
 		</>
 	);
 }
